Rework of state model - remove SLEEPING class. That's a transition state that cannot be learned properly - remove subsidiary feature (linked to responsible) - change classifier (logistic) param. Use L2 penalty and higher C1 to smooth %proba. Use auto class weight to better predict non-WON leads - change proba from float to int - display top 10 coef of each class in model eval function

+++ leads/models.py
@@ -258,7 +258,7 @@ class StateProba(models.Model):
     """Lead state probability"""
     lead = models.ForeignKey(Lead)
     state = models.CharField(_("State"), max_length=30, choices=Lead.STATES)
-    score = models.FloatField(_("Score"))
+    score = models.IntegerField(_("Score"))
 
 
 # Signal handling to throw actionset and document tree creation
